Moved _grid-settings to Neat
Move _grid-settings.scss from Bitters to Neat. It is now imported right
after _neat-helpers but before all other Neat scss files without user
intervention. This would seem to remove concerns about the order in
which Bitters and Neat are imported after Bourbon, and, I believe,
remove the necessity of having a different path to neat-helpers when
running in (or not running in…) Rails. That said, this all seems a
little too obvious so I feel like I’m missing the reason it wasn’t done
initially.